Adding access to certificate bundles. porting the changes from r2.21.x
Closes-Bug: #1494892
(cherry picked from commit 119df693d5edc0544c33d7802be38ff3582d0b75)

+++ src/config/common/utils.py
@@ -21,6 +21,7 @@
 # @author: Numan Siddique, eNovance.
 
 
+import os
 import re
 import urllib
 from collections import OrderedDict
@@ -194,10 +195,13 @@ def obj_type_to_vnc_class(obj_type, module_name):
 
 
 def getCertKeyCaBundle(bundle, certs):
+    if os.path.isfile(bundle):
+        return bundle
     with open(bundle, 'w') as ofile:
         for cert in certs:
             with open(cert) as ifile:
                 for line in ifile:
                     ofile.write(line)
+    os.chmod(bundle,0o777)
     return bundle
 # end CreateCertKeyCaBundle
